Bug fix. Can now enter decimal type for distance

+++ Freewire/Controllers/ToolsController.cs
@@ -25,7 +25,7 @@ namespace Freewire.Controllers
         // Called when submitting form data
         // GET: Tools
         [HttpPost]
-        public ActionResult RadioQualifier(int distance, int bandwidth, string sort, string lastSort, string lastSortOrder)
+        public ActionResult RadioQualifier(double distance, int bandwidth, string sort, string lastSort, string lastSortOrder)
         {
             //Keeps previously entered data in the form fields.
             ViewBag.distance = distance;
